refactor main.py to /src/models/api/app.py
